Changed project layout/structure
Removed un-needed namespaces (No need for a separate namespace per
class/script)

Added brackets to if statement. This is a small thing, but unless it is
a trivial if statement, probably best to leave the brackets in there.
Even if it is 1 line of code spaced over 2.

Fixed movement so that it does not make the player move faster
diagonally.

Added private to the variables. This isn't needed obviously but i
personally think it is a good habit to get into. If people disagree we
can remove them?