Updated the Movement and Player scripts
Changed the way the player moves. Instead of directly moving from the
player script, it tells the movement script to move the character. This
decouples movement from the player. This means (for example) that the
player script could be replaced with an AI script and everything would
function the same.

A template was added for the C# scripts. Use this so that you don't have
to add the namespace every time.