Added basic Ability creation, UI elements and test script.
I added the base class for abilities and created a dummy ability called
'Slash' for testing purposes (It does nothing just logs to the console
when it is used).

I also worked on a crude implementation of the UI. 4 'abilities' to the
bottom of the screen, button that has an active ability can be used. So
far only ability 1 is in use (Slash).

There is also a health bar in the top left. Just simple green bar that
moves up and down as the players health does.

Lastly a test script was added so that new features could easily be
tested. For now this add's the Slash ability to ability slot 1, and
constantly heals and damages the player.