Added basic stance changing/UI
Added a basic form of stances. Played around with an idea for how to
display the stances, but still stupidly early prototype. Buttons work
and change stances, but don't do anything except log some stuff to the
console.

I'll also get around to cleaning the code up for the stances, it's not
very good atm