More Cleanup
Added _old suffixes to (hopefully) everything in scripts that needs to
be re-implemented in some form or looked at, then consolidated them in a
_old folder.

This is just so that we can still look at how it was done before and
have the relevant code to copy-paste but we know what we need to rebuild
overall.

Got rid of our script folders other than Stances and _old until we have
enough to justify again.